Addressing suggestions for PR:  - Fixed "simplificiation" typo in README    + Also changed "simplification_interpolation" in README to      "simplification_interpolate" to reflect correct argument    + Arguments not beginning in "simplification_" were being skipped in      configure_robot_and_planner_with_kwargs -- fixed by changing a "continue" into an if statement  - Changed "fcit_single"/"fcit" to "fcit"/"fcit_multi_goal"  - Moved "QueueEdge" out of utils.hh and into fcit.hh    + Should we also move "FCITRoadmapNode" out of roadmap.hh as well?      It's only used by FCIT*  - Changed "(*it)." to "it->"  - Changed invalidList to an unordered_set (best results)
